Refactoring: remove all custom marker logic from Map
